==> 02_Scripts/1_DataImport_BasicStatisticalAnalysis.ipynb <==
# Importing data sets and conducting basic statistical analysis

Understanding the size and shape of our dataset 

1. Importing and basic statistical analysis
    1. Importing libraries
    2. Importing data sets
    3. Checking the size, shape, and content of the datasets
    
2. Data Wrangling
    1. Dropping columns
    3. Renaming columns
    
3. Importing new departments dataset and transposing for future analysis 
    1. Importing dataset
    2. Transposing
    3. Changing the header
    4. Data dictionaries
    5. Testing out subsetting on new data dictionary using loc and isin functions

4. Exporting datasets

# 1. Importing and basic statistical analysis

### A. Importing libraries

# import libraries
import pandas as pd
import numpy as np
import os

### B. Importing datasets

#folder path into usable string
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import orders dataset
df_ords = pd.read_csv(os.path.join(path, '02_data', 'Original_data', 'orders.csv'), index_col = False)

#import products dataset
df_prods = pd.read_csv(os.path.join(path, '02_data', 'Original_data', 'products.csv'), index_col = False)

### C. Checking size, shape, and content of datasets 

df_ords.head()

df_ords.head(30)

df_ords.shape

df_ords.info()

df_ords.dtypes

df_ords.columns

df_ords.dtypes

df_ords.describe()

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.info()

df_prods.describe()

**Observations**

The dataset, *orders*, is comprised of seven columns and 3,421,083 rows.  
The dataset, *products*, is comprised of five columns and 49,693 rows. 

Initial observations show that there could be an error in the prices of the products dataset since the maximum price is **99,999**. 

# 2. Data Wrangling

### A. Dropping columns

The columns *eval_set* in the orders dataset is not needed for analysis so will be removed from the dataset. 

#dropping eval_set column from orders.csv 
df_ords.drop(columns = ['eval_set'])

#defining df_ords without the eval_set column
df_ords = df_ords.drop(columns = ['eval_set'])

#checking output
df_ords.head(5)

### B. Renaming columns

Renaming *order_dow* to *orders_day_of_week* to be clearer

#renaming columns
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

#checing output
df_ords.head(5)

# 3. Importing departments dataset

We need the departments dataset so we can cross reference the department id with department name. This will help later in our analysis when we want to know which departments account for the most spends. 

### A. Importing dataset

#importing data set departments.csv
df_dep = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'departments.csv'), index_col = False)

### B. Transposing dataset

#checking output of dataset
df_dep.head()

#transposing dataset for future merging
df_dep.T

#saving newly transposed dataset as df_dept_t
df_dep_t = df_dep.T

#checking output 
df_dep_t

#adding an index
df_dep_t.reset_index()

### C. Changing the header

When we transposed the data, the header row became row 0. We now need to reset the header. 

#take the first row of the df_dep_t for the header
new_header = df_dep_t.iloc[0]

#checking output
new_header

#take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

#return the new df
df_dep_t_new

#set the header row as the df header
df_dep_t_new.columns = new_header

#checking output
df_dep_t_new

### D. Data dictionaries

We need to create a data dictionary of the department ids and department names so we can use this in our future analysis. 

#creating data dictionary
data_dict = df_dep_t_new.to_dict('index')

#checking output
data_dict

#using the data dictionary to understand values in the data frame
df_prods.head()

#checking output of one department
print(data_dict.get('19'))

### E. Testing out subsetting on new data dictionary 

Testing out creating subsets of data dictionary using loc and isin functions. 

#using the loc function to subsetting
df_snacks_2 = df_prods.loc[df_prods['department_id']==19]

df_snacks_2.head()

#using loc combined with isin to search multiple values
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([17, 18, 19])]

df_snacks_3

# 4. Export datasets

#export df_ords as orders_wrangled.csv
df_ords.to_csv(os.path.join(path, '02_Data', 'Prepared_data', 'orders_wrangled.csv'))

#export the df_dep_t_new dataframe as departments_wrangled.csv
df_dep_t_new.to_csv(os.path.join(path, '02_Data', 'Prepared_data', 'departments_wrangled.csv'))
