==> 02_Scripts/2_DataConsistency.ipynb <==
# Data Consistency Checks

1. Importing libraries and datasets
2. Investigating mixed-type columns
3. Duplicates
4. Additional exploratory analysis of dataset
    1. Describe
    2. Mixed-type data
    3. Missing values
    4. Duplicate values
5. Export dataframes

# 1. Importing libraries and datasets

#import libraries
import pandas as pd
import numpy as nm
import os

#folder path into usable string
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import products csv
df_prods = pd.read_csv(os.path.join(path, '02_data', 'Original_data', 'products.csv'), index_col = False)

df_prods

#import orders_wrangled csv
df_ords = pd.read_csv(os.path.join(path, '02_data', 'Prepared_data', 'orders_wrangled.csv'), index_col = False)

df_ords

#dropping unnamed column
df_ords.drop(columns = ['Unnamed: 0'])

#defining df_ords without the eval_set column
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

#checking output
df_ords

# 2. Investigating mixed-type columns

Mixed-type columns are where you have different data types within the same column, e.g., string and interger. It's important that we clean mixed-data *before* we conduct our analysis, as it will interfer with our code. 

*Since our dataset doesn't contain any mixed-data columns, we need to create one for the purposes of this exercise.*

#create a dataframe
df_test = pd.DataFrame()

#create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

#calculating the number of null values in your df
df_prods.isnull().sum()

#creating df of just null values  
df_nan = df_prods[df_prods['product_name'].isnull()==True]

df_nan

df_prods.shape

#create new clean df without missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

#run shape() function again to compare the rows and columns in your new df
df_prods_clean.shape

# 3. Duplicates

We need to check for duplicate values, in this instance we're checking for duplicate rows rather than duplicate values in columns since it is expected that we would have duplicate in say, the department id column. 

#looking for full duplicates by creating a new df
df_dups = df_prods_clean[df_prods_clean.duplicated()]

#returning all duplicate values
df_dups

**Observations**

Five duplicates were found, so we will remove these from our dataset. 

#checking the shape of my df so I can compare once removed duplicates
df_prods_clean.shape

#create a df that doesn't include duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#checking output
df_prods_clean_no_dups.shape

# 4. Additional exploratory analysis of dataset

I want to run some additional basic statistical analysis to further understand our dataset. Initially we're looking at whether the minimum, maximum, and mean values in our dataset are consistent with our expectations. 

### A. Describe function

#run the df.describe function on df_ords
df_ords.describe()

**Observations**

Broadly speaking, the values are what I would expect. The maximum value in order_id is 3.421083e+06 which aligns with the number of rows in our dataset, as do the total number of orders in the order_number column. 
Fot order_hour_of_day, our minimum value is 0 and our maximum is 23 which aligns with there being 24 hours in the day. 

### B. Mixed-type data

#Q3 Check for mixed-type data in your dataframe
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print (col)

**Observations**

No mixed data returned

### C. Missing values

#Q5 run a check for missing values in your dataframe
##firstly to see Ture/False
df_ords.isnull()

##secondly now grouped 
df_ords.isnull().sum()

**Observations**

There are 206,209 null values in a dataframe of 3,421,083 rows = 6% of the column values. My initial thought would be that these could be first-time purchasers. A null values in the days_since_prior_order column would be expected because they have never ordered before. 

My response to the 206,209 missing values in the prior order column would be to leave them as is. 
Firstly, 6% of values is a reasonably small amount of the column to be missing, so keeping these values as null wouldn't be statistically significant. I don't think imputing values based on the mean would bring much benefit to my analysis.
Secondly, I think it an interesting datapoint which the client may benefit from understanding. We could create a flag which highlights users with no prior orders, ie., new customers. Assuming my assumption of the reason behind the null values is correct, I would say this client has high customer retention with a lot of repeat purchasers, but perhaps they should be doing more to bring in new customers as currently they only form 6% of their customer base, in the time period selected. 

### D. Duplicate values

#run a check for duplicate values in your df_ords data
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

#double checking using the shape() function 
df_ords_dups.shape

**Observations**

There are no duplicate rows but if there were, I would have created a new dataframe: df_ords_no_dups = df_ords.drop_duplicates()

# 5. Export data frames

#export cleaned df_prods as csv
df_prods_clean_no_dups.to_csv(os.path.join(path, '02_Data', 'Prepared_data', 'products_clean.csv'))

#export cleaned df_ords as csv
df_ords.to_csv(os.path.join(path, '02_Data', 'Prepared_data', 'orders_clean.csv'))
