==> 02_Scripts/3_MergingDataSets.ipynb <==
# Merging products and order dataframes

Combining orders data set with new data, order_products_prior

1. Import libraries and data set
2. Removing unwanted indexes
3. Merging data set 
4. Combine newly merged dataframe with products dataset
5. Exporting data frame

# 1. Import libraries and data set

#import libraries
import pandas as pd
import numpy as nm
import os

#folder path into usable string
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import order_products_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'order_products_prior.csv'))

#check output
df_ords_prior.head()

#import orders
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_data', 'orders_clean.csv'))

#checking output 
df_ords

#import prods data set
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_data', 'products_clean.csv'),index_col=False)

#checking output
df_prods

#dropping unnamed column
df_prods.drop(columns = ['Unnamed: 0'])

#defining df_ords without the eval_set column
df_prods = df_prods.drop(columns = ['Unnamed: 0'])

#checking output
df_prods

# 2. Removing unwanted indexes

An unwanted column, *Unnamed: 0*, has appeared in the dataset so we will remove this before we conduct our analysis

df_ords.head()

#removing column called 'Unnamed 0'
df_ords = df_ords.drop(['Unnamed: 0'],axis=1)

#checking output
df_ords.head()

#checking shape
df_ords_prior.shape

#checking shape
df_ords.shape

# 3. Merging orders dataset with prior orders dataset

We are going to merge our clean dataset, orders, with a new dataset called *prior orders* which will give us information on the product id. Our join will be made on the shared variable, *order_id*. 

#merging df_ords and df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

#checking output
df_merged_large.head()

#checking output
df_merged_large.shape

df_merged_large['_merge'].value_counts()

df_merged_large.describe()

**Observations**

Our new dataframe has the same number of rows as before, so the merge is successful. 

# 4. Combine newly merged dataframe with products dataset

Merging with the products dataset will give us crucial information about the products purchased in each order. Using the merge() function on the common variable, 'product_id' will enable us to see any instances where the product_id doesn't match, as they would appear as NaN values. 

#merging data sets
df_merged_final = df_merged_large.merge(df_prods, on = ['product_id'])

df_merged_final.head()

#check output
df_merged_final['_merge'].value_counts()

**Observations**

The dataset has the same number of rows, so the merge was successful.

# 5. Export data frame

#export new dataframe in pickle 
df_merged_final.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'full_merged.pkl'))
