==> 02_Scripts/4_DerivingNewVariables.ipynb <==
# Deriving New Variables Exercise

1. Import libraries and data set
2. Creating price label column
3. Creating busiest day column
4. Export dataset 

## 1. Import libraries and data set

#import libraries
import pandas as pd
import numpy as nm
import os

#create path to data source
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import merged dataframe
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'full_merged.pkl'))

## 2. Creating price label column

This new price label column will group our products into one of three price ranges: high, mid, or low. 

#using the loc() function to create price label column
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low_range product'

df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

**Observations**

The majority of products fall within the mid price range, between 5 and 15. The next most popular price range is low, anything up to $5. Finally, high range products account for the least amount of products. 

## 3. Creating busiest day column

The client would like to know which days are the most popular for purchases. 

Days of the week codes:
    0 = Saturday,
    1 = Sunday,
    2 = Monday,
    3 = Tuesday,
    4 = Wednesday,
    5 = Thursday, 
    6 = Friday 

#calculate which days have the most orders
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

#create list for new busiest day column
result = []

for value in df_ords_prods_merged['orders_day_of_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

result

#add the new column to your data set
df_ords_prods_merged['busiest_day'] = result

#print the frequency of the new column to check with the frequencies stated before
df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

**Observations**

Saturday and Sunday are the busiest days of the week, followed by Friday, Monday, and Thursday, and finally Tuesday and Wednesday are the least busy.

## 4. Creating busiest days column

The clients would like to create a 'busiest days' (plural) column. This label should correspond with the two busiest days of the week as opposed to the single busiest day. They'd also like to know the two slowest days. 

Days of the week codes
    0 = Saturday,
    1 = Sunday,
    2 = Monday,
    3 = Tuesday,
    4 = Wednesday,
    5 = Thursday, 
    6 = Friday 

#calculate which days have the most orders
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

Two busiest days are 0 and 1 (Saturday and Sunday) and the two slowest days are 3 and 4 (Tuesday and Wednesday)

#create list for new busiest days column
result = []

for value in df_ords_prods_merged['orders_day_of_week']:
    if value == 0 or value == 1:
        result.append('Busiest days')
    elif value == 3 or value == 4:
        result.append('Slowest days')
    else:
        result.append('Regularly busy')

result

#add the new column to your data set
df_ords_prods_merged['busiest_days'] = result

#print the frequency of the new column to check with the frequencies stated before
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

**Observations**

Having compared the results against those in line The main observation is that the two busiest days, Saturday and Sunday, account for over a third of sale days - 37%. The next three most popular days, Monday, Thursday, and Friday, make up the largest category, "Regularly busy", but only 3% more sales for a total of 40%. Finally, the two slowest day, Tuesday and Wednesday, account for 23% of sales. 

# 5. Creating busiest hour of day column

When too many users make Instacart orders at the same time, the app freezes. We need to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled "Most orders", "Average orders", and 'Fewest orders'. 

#check to see which column we need, 'order_hour_of_day'
df_ords_prods_merged.head()

#print the frequency of order_hour_of_day
df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

**Observations:** Based on the above values, I am assigning anything with a frequency of 2.4m or higher as "Most orders" (9:00 - 16:00), anything above 100,000 up to 2.4m as "Average orders" (7:00 - 8:00, and 17:00 - 22:00), and anything under 100,000 as "Fewest orders" (23:00 - 6:00). 

#create list for new busiest period of day column
result = []

for hour in df_ords_prods_merged['order_hour_of_day']:
    if hour in [10, 11, 14, 15, 13, 12, 16, 9]:
        result.append('Most orders')
    elif hour in [23, 6, 0, 1, 5, 2, 4, 3]:
        result.append('Fewest orders')
    else:
        result.append('Average orders')

result

#add the new column to your data set
df_ords_prods_merged['busiest_period_of_day'] = result

#print the frequency of the new column 
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

#checking my columns names 
df_ords_prods_merged.head()

#final check of dataframe
df_ords_prods_merged.head()

# 4. Export dataset

#export dataframe 
df_ords_prods_merged.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'merged_new_variables.pkl'))
