==> 02_Scripts/5_AggregatingData.ipynb <==
# Grouping Data and Aggregating Variables 

1. Import libraries and data set 
2. Aggregating data with agg() function
    1. Derive a new column, loyalty flag
    2. Create a new column, loyalty flag, with transform() and loc()
4. Looking at spending habits of different customers
    1. Checking basic statistics
    2. Looking at loyalty flag as percentage of whole
    3. Investigating outlier values
5. Creating a spending flag
    1. Create spending flag column
    2. Assign users based on spending habits 
6. Creating an order frequency flag 
    1. Create order freq flag column
    2. Assign users based on order frequency habits 

## 1. Import libraries and data set

# import libraties
import pandas as pd
import numpy as np
import os

#create usable path 
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import ords_prods_merge dataset
df = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'full_merged4.pkl'))

#check dataframe output
df.head()

#drop max_order and loyalty flags columns since I will be recreating them as part of this submission
df = df.drop(['max_order', 'loyalty_flag'], axis = 1)

#checking out put 
print('Shape of new dataframe:', df.shape)

print('Sample of new dataframe')
df.sample(5)

## 2. Aggregating data with the agg() function

Aggregating the mean of the "order_number" column grouped by "department_id" for the entire dataframe. This will give us the minimum, maximum, and mean of orders by department. 

# group df_ords_prods_merge by the 'department_id' column, then calculate the mean, min and max of the 'order_number' column
print('The mean, min and max of order_number grouped by the department_id column in df:')
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

#### Creating a subset to compare aggregations against the entire dataframe

#creating subset and printing to check output
df_subset = df[:1000000]

print('First five results of df_ords_prods_merge subset')
df_subset.head(5)

# group df_subset by the 'department_id' column, then calculate the mean, min and max of the 'order_number' column
print('The mean, min and max of order_number grouped by the department_id column in df_subset:')
df_subset.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

**Observations:** 

Whilst there are some differences between the aggregations seen in the subset compared to the entire dataframe, they tend to be within reasonable ranges. I calculated the percentage change for four departments: department 1 was within 4%, department 8 was within 7%, department was within 2%, and department 21 was within 4%. 

There is some variance in the max values in the subset. The max values for all departments in the entire dataset is 99, however, in some of the subset department it is lower, for example, department 8 is only 91. Again, this is not unreasonable to expect when only looking at a slice of the data. 

## 3. Creating a loyalty flag for customers using transform() and loc() functions

The client would like to better understand the loyalty status of their customer base, i.e., the extent to which customers return to purchase from Instacart. We are going to column which calculates the maximum number of orders per user, and then create a flag which groups users into one of three categories based on how many orders they have made. 

### A. Creating a new column, 'max_order' to place the results of the maximum orders for each user

# create a new column 'max_order2' containing the maximum order number for each user_id, calculated using the groupby() and transform() functions
df['max_order'] = df.groupby(['user_id'])['order_number'].transform(np.max)

#check output by looks at user_id and max_order
print('Sample user_id and max_order output:')
df[['user_id', 'max_order']].sample(10)

df.sample(5)

### B. Derive a new column, 'loyalty_flag' using the loc() function and the values found in "max_order"

Creating a flag which will assign each customer into one of three categores based on how many times they have purchased from Instacart. This loyalty flag allows us to group our customers by their loyalty status. 

df.loc[df['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df.loc[df['max_order'] <= 40, 'loyalty_flag'] = 'Regular customer'

df.loc[df['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#checking ouput of flag creation
df['loyalty_flag'].value_counts(dropna = False)

# checking the accuracy of the 'loyalty_flag' column with output from columns 'user_id', 'max_order', and 'loyalty_flag' column
print('Sample output from columns user_id, max_order and loyalty_flag columns:')
df[['user_id', 'max_order', 'loyalty_flag']].sample(20)

## 4. Looking at the spending habits of different customers

Looking at how loyalty status impacts spending habits. 

### A. Check basic statistics of product prices for each loyalty category (loyal, regular, and new), and assess whether spending habits differ across the different groups. 

#basic statistical comparison of 'loyalty_flag' column, followed by calculation of mean, min, max, and sum of the 'prices' column
print('The mean, min, max, and sum of prices grouped by the loyalty_flag:')

df.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max', 'sum']})

#confirming the numbers of each type of customer
df.groupby('loyalty_flag')['user_id'].count()

### B. Calculating each loyalty flag as a percentage of total orders

#setting loyalty_flag as variable 
loyalty_flag = df.loyalty_flag

#calculating loyalty_flag as % of whole
percent_loyalty_flag = loyalty_flag.value_counts(normalize = True)
print(percent_loyalty_flag)

#returning results in easier to read format 
percent_loyalty_flag_100 = loyalty_flag.value_counts(normalize = True).mul(100).round(1).astype(str) + '%'
print(percent_loyalty_flag_100)

**Observations**

Almost 50% of orders fall within the regular customer category, just over 30% are from loyal customers, and just under 20% are new customers. Interestingly, the average price for loyal customers is the lowest of the three groups whilst new customers have the highest which I was surprised by. Some interesting futher analysis could be focused on the types of products loyal customers are buying.    

The maximum price value of $99,999 seems like an extreme outlier, and requires further investigation. Significant outliers like this could skew our data.  

### C. Investigating the outlier prices

#returning a list of all prices, ordered by most expensive
df['prices'].drop_duplicates().sort_values(ascending = False)

There are two prices which appear to be outliers: 99,999 and 14,900

#filtering the dataset by two outlier prices
max_price = df.loc[df['prices'].isin([99999, 14900])]

max_price.sample(5)

max_price.shape

**Observations**

There are 5127 rows impacted by these outlier prices of 99,999 and  14,900. Ideally I would conduct some more investigation to see if we can figure out the correct price and amend the dataset to reflect. Results including these values as they currently are may impact the mean values returned in the basic statistics returned above. 

## 5. Create a spending flag for each user based on the average price across all their order

The client would like to better understand how customers are spending money.  
* If the mean of the prices of products purchased by a user is lower than 10, then flag them as a "Low spender".
* If the mean of the prices of products purchased by a user is higher than or equsl to 10, then flag them as a "High spender"

### A. Create new column called spending flag

#create a new columns called "spending_flag"
df['avg_price'] = df.groupby(['user_id'])['prices'].transform(np.mean)

#round the averages
df['avg_price'] = df.groupby(['user_id'])['prices'].transform(np.mean).round(2)

#check output
print('Sample of user_id and average price output:')
df[['user_id', 'avg_price']].sample(10)

### B. Create a flag which tags any user with an average spend of more than or equal to 10, as "high spender". Any customers with an average spend of less than 10 will be marked as "Low spender"

df.loc[df['avg_price'] >= 10, 'spending_flag']='High spender'

df.loc[df['avg_price'] < 10, 'spending_flag']='Low spender'

#check output
df.sample(5)

## 6. Create an order frequency flag which corresponds to the regularity of a customer's order behaviour

The client would like to better understand the frequency of customer purchases
* If the median of "days_since_prior_order" is **higher than 20**, the customer should be labeled a **"Non-frequent customer"**
* If the median of "days_since_prior_order" is **higher than 10 and lower than or equal to 20**, then the customer should be labeled a **"Regular customer"**
* If the median is **lower than or equal to 10**, then the customer should be labeled a **"Frequent customer"**

#create a new column "order_freq" containing the median in days_since_prior_order
df['order_freq'] = df.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# check the output of 'user_id' and 'order_freq'
print('Sample user_id and order_freq output:')
df[['user_id', 'order_freq']].sample(10)

#### Create a flag which tags any user with a median frequency order of less than or equal to 10 as "non-frequent customer", higher than 10 and lower than or equal to 20 as "regular customer", and higher than 20 as "frequent customer"

# create the 'order_freq_flag' column based on each user's median in days_since_prior_order
df.loc[df['order_freq'] > 20, 'order_freq_flag'] = 'Non-frequent customer'

df.loc[(df['order_freq'] > 10) & (df['order_freq'] <= 20), 'order_freq_flag'] = 'Regular customer'

df.loc[df['order_freq'] <= 10, 'order_freq_flag'] = 'Frequent customer'

#check the output and count of each new flag

print('Count of each order frequency variable')
df['order_freq_flag'].value_counts(dropna = False)

# Export Dataset

df.to_pickle(os.path.join(path, '02_Data','Prepared_data', 'full_merged5.pkl'))
