==> 02_Scripts/6_CleaningNewDataset_Customers.ipynb <==
# Creating charts and visualisations in Python

1. Importing libraries and datasets 
2. Addressing our new data set
    1. Wrangling 
    2. Cleaning
    3. Merging with existing dataset 
3. Creating visualisations (included in Part2 notebook)

# 1. Importing libraries and datasets

#import libraries
import pandas as pd
import numpy as np
import os

#create path
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import customers data set
df = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'customers.csv'), index_col=False)

#checking output
print('Sample of new customers data set')
df.head(10)

print('Number of rows and columns in customers data set')
df.shape

##### Initial observations:

Our new customers data set provides personal information about Instacart's customer base including first name, surname, gender, state location, age, when they joined Instacart, the number of dependents they have, their family status, and finally their income. Already there are some spelling and syntactical errors in the column names which we will be able to fix to make the data set easier to understand. 

There are 206,209 rows, so lot's of customers! 

# 2. Addressing our dataset

### A. Data Wrangling

#### Approach

1. Some simple exploratory analysis to understand a couple of the columns.
2. Removal of the first_name, and surname columns. User_id is a more useful customer identifier than first or surname.  
3. Amend any spelling and syntactical errors in the column titles. This will include use of capital and lowercase letters, and use of underscores and spaces. 
4. Check data types of columns to ensure they are correct and appropriate for the type of data collected
5. Use a data dictionary to change the data within the STATE column to two-letter codes. Although not specifically needed, reducing the data to two-letters is cleaner, and would make it easier to, for example, search for specific states using the isin fucntion. 
6. Create a new column with flag related to whether or not the customer has dependents. Instacart could provide targeted advertising to individuals based on whether they have dependents on not.


##### 1. Initial exploratory analysis to understand some of the columns

#checking gender column
df['Gender'].value_counts()

#checking age column
df['Age'].value_counts()

df['Age'].describe()

#checking fam_status column
df['fam_status'].value_counts()

#checking income column
df['income'].describe()

#checking n_dependants
df['n_dependants'].describe()

##### Observations:

I wanted to understand more about the types of values contained within these columns before I started amending anything. Everything so far seems reasonable. 

##### 2. Removal of first name and surname columns

#remove date_joined column using drop function
df.drop(columns = ['First Name', 'Surnam'])

#define dateframe minus date_joined column
df = df.drop(columns = ['First Name', 'Surnam'])

#check output
df.head()

#checking shape
df.shape

##### 3. Amending spelling and syntax errors in columns names

#amending First Name, Surnam, Gender, STATE, and Age columns
df.rename(columns = {'Gender':'gender', 'STATE':'state', 'Age':'age'}, inplace = True)

df.head()

##### 4. Checking datatypes

#returning the datatypes for all columns
df.info()

###### Observations:

Whilst the majority of the datatypes themselves are all correct, we should reduce the size of some of them to save on space. The date_joined column should be changed to datetime. 

#amending datatype for user_id, age, n_dependents, and income
df['user_id']=df['user_id'].astype('int32')
df['age'] =df['age'].astype('int8')
df['date_joined'] = df['date_joined'].apply(pd.to_datetime)
df['n_dependants']=df['n_dependants'].astype('int16')
df['income']=df['income'].astype('int32')

#checking output
df.info()

df.head()

##### 5. Using a data dictionary to change state names to two-letter abbreviations

#creating list of all state names
states_listed = df['state'].unique()

#checking output
print(sorted(states_listed))

#number of each state
df['state'].value_counts().sort_index()

#counting number of states
state_count = df['state'].nunique()
print(state_count)

#creating data dictionary, abbreviations were retrieved from US government website
state_dict = {'Alabama': 'AL',
              'Alaska': 'AK', 
              'Arizona': 'AZ',
              'Arkansas': 'AR',
              'California': 'CA',
              'Colorado': 'CO',
              'Connecticut': 'CT',
              'Delaware': 'DE',
              'District of Columbia': 'DC',
              'Florida': 'FL',
              'Georgia': 'GA',
              'Hawaii': 'HI',
              'Idaho': 'ID',
              'Illinois': 'IL',
              'Indiana': 'IN',
              'Iowa': 'IA',
              'Kansas': 'KS',
              'Kentucky': 'KY',
              'Louisiana': 'LA',
              'Maine': 'ME',
              'Maryland': 'MD',
              'Massachusetts': 'MA',
              'Michigan': 'MI',
              'Minnesota': 'MN',
              'Mississippi': 'MS',
              'Missouri': 'MO',
              'Montana': 'MT',
              'Nebraska': 'NE',
              'Nevada': 'NV',
              'New Hampshire': 'NH',
              'New Jersey': 'NJ',
              'New Mexico': 'NM',
              'New York': 'NY',
              'North Carolina': 'NC',
              'North Dakota': 'ND',
              'Ohio': 'OH',
              'Oklahoma': 'OK',
              'Oregon': 'OR',
              'Pennsylvania': 'PA',
              'Rhode Island': 'RI',
              'South Carolina': 'SC',
              'South Dakota': 'SD',
              'Tennessee': 'TN',
              'Texas': 'TX',
              'Utah': 'UT',
              'Vermont': 'VT',
              'Virginia': 'VA',
              'Washington': 'WA',
              'West Virginia': 'WV',
              'Wisconsin': 'WI',
              'Wyoming': 'WY'}

#replace values in state column with state dictionary 
df2 = df.replace({'state': state_dict})

df2.sample(10)

df2['state'].value_counts().sort_index()

The value counts for the new column matches those for the state column in the original dataframe

##### 6. Create a new flag column on whether the customer has dependants 

df2.loc[df2['n_dependants'] >= 1, 'dependants_loc'] = 'Has dependants'

df2.loc[df2['n_dependants'] < 1, 'dependants_loc'] = 'No dependants'

df2['dependants_loc'].value_counts(dropna = False)

df2.head()

Although the code has worked, I would prefer for my two columns related to dependents to be situated next to one another. 

#creating a list of my column titles
titles = list(df2.columns)
titles

#creating a new list, with the new changed order of columns
new_titles = ['user_id',
 'gender',
 'state',
 'age',
 'date_joined',
 'n_dependants',
 'dependants_loc',
 'fam_status',
 'income']

#assigning my new column title order to my dataframe
df3 = df2[new_titles]

#checking my output
df3.sample(5)

#### Final check of shape and size of dataframe 

df3.sample(10)

df3.shape

##### Final Observations

In terms of the shape of our dataframe, we have the same number of rows as before. We have dropped three columns, first name, surname, and date_joined, and created another, dependants_loc. 

We have also ensured consistency in our column naming conventions and changed the state names to abbreviations. 

### B. Data Quality and Consistency 

#### Approach 

We're going to be looking at a few things in terms of data quality and consistency 

1. Checking for null values
2. Checking for duplicates
3. Checking for mixed-type data 

1. Checking for null values

df3.isnull().sum()

##### Observations:

No missing data observed

2. Checking for duplicates

df_dups = df3[df3.duplicated()]

df_dups

##### Observations:

No duplicate rows found

3. Checking for mixed-type data 

for col in df3.columns.tolist():
    weird = (df3[[col]].applymap(type) != df3[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df3[weird]) > 0:
        print (col)

##### Observations:

No mixed-type data identified. 

Interestingly, when I ran this code on the original data there were over 11,000 mixed-type data points in the first name field, representing around 5.5.% of the data set, but given that name is not important to our statistical analysis we have removed the column. 

## C. Merging

Merging our new customers dataset with our existing orders and products dataset

#import orders and products dataset 
df_merge = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'full_merged5.pkl'))

#check output
df_merge.head()

df_merge.shape

df_merge.describe()

#reminder of our new data set to be merged 
df3.head()

df3.shape

To merge our two dataframes, we need to have a common variable. In this case, it is user_id.

#merging our datasets
df_combo = df_merge.merge(df3, on = ['user_id'])

df_combo.sample(5)

df_combo.shape

#Checking value counts
df_combo['_merge'].value_counts()

df_combo.describe()

#checking on space constraints
df_combo.info()

## Export dataframe

#export to pickle 
df_combo.to_pickle(os.path.join(path, '02_Data','Prepared_data', 'orders_products_all.pkl'))
