==> 02_Scripts/8_CustomerAnalysis_GeographicActivityFlags.ipynb <==
# Final Project 

#### Please note, due to memory constraints we are using a 70% randomised and representative sample of the full data set for our analyses and visualisations. This subet was created in a previous exercise. 

This notebook focused on creating additional new variables and dataframes, to futher our analysis. 

###  Contents:

1. Importing libraries and files
2. Considering security implications of our data
3. Comparing customer behavior in different geographic areas across the US
    1. Assigning states to geographic regions
    2. Analysing spending by regions
4. Creating an exclusion flag for low-activity customers (customers with less than 5 orders) and excluding them from the data. 
    1. Low activity dataframe
    2. High activity dataframe
5. Exporting new data frames
    
   ** **

# 1. Importing Libraries and Files

#import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

#create usable path 
path = r'C:\Users\rutha\CareerFoundry\01-23_Instacart_Basket_Analysis'

#import data set
df = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'big_sample.pkl'))

#checking output
print ('A sample of our orders and products data:')
df.head()

df.shape

df.info()

## 2. Considering security implications in the data

Looking at whether we have any personal identification information in our dataset, which we would need to flag. 

df.head()

df.info()

**Observations**

In our current merged data set we have no personal identification information (PPI). 

In one of our original data sets, 'customers', we had two columns, first name and surname, which would have fallen within the parameters of PPI. However, given that we have the user_id information, a better unique identifier, we decided to drop those two columns. 

## 3. Comparing customer behaviour across different regions

Look at regional segmentation of the dataset, we will create a “Region” column based on the “State” column from the customers data set.**

**Determine whether there’s a difference in spending habits between the different U.S. regions.**

### 3.1 Assigning states to geographic regions

# creating a new "region" column
df['state'].value_counts().sort_index()

#defining region lists 
region1_NE = ['ME', 'NH', 'VT', 'MA', 'RI', 'CT', 'NY', 'PA', 'NJ']
region2_MW = ['WI', 'MI', 'IL', 'IN', 'OH', 'ND', 'SD', 'NE', 'KS', 'MN', 'IA', 'MO']
region3_S = ['DE', 'MD', 'DC', 'VA', 'WV', 'NC', 'SC', 'GA', 'FL', 'KY', 'TN', 'MS', 'AL', 'OK', 'TX', 'AR', 'LA']
region4_W = ['ID', 'MT', 'WY', 'NV', 'UT', 'CO', 'AZ', 'NM', 'AK', 'WA', 'OR', 'CA', 'HI']

#assigning regions values to a new column (region) based on information from US Census Bureau

df.loc[df['state'].isin(region1_NE), 'region'] = 'Northeast'
df.loc[df['state'].isin(region2_MW), 'region'] = 'Midwest'
df.loc[df['state'].isin(region3_S), 'region'] = 'South'
df.loc[df['state'].isin(region4_W), 'region'] = 'West'

#checking regions output
df['region'].value_counts(dropna = False)

### 3.2 Analysing spending by regions

#creating a spending_habits crosstab
crosstab_spend = pd.crosstab (df['region'], df['spending_flag'], dropna = False)

crosstab_spend

**Observations**

It appears that the South accounts for the largest proportion of purchases, and the Northeast for the least. We will need to cross check these against population data to see if that aligns. 

## 4. Customers with Low Activity

The client asked us to exclude any customers with fewer than 5 purchases. So we will create an exclusion flag for customers with low-activity (customers with fewer than 5 orders), and exclude them from the data

# creating the exclusion flag 
df.loc[df['max_order'] < 5, 'exclusion_flag'] = 'low-activity'

df.loc[df['max_order'] >= 5, 'exclusion_flag'] = 'high-activity'

#checking new added column
df['exclusion_flag'].value_counts(dropna = False)

df.info()

### 4.1 Excluding low-activity customers from the data and export the subset.

#creating a new dataframe minus low-activity customers
df_high = df.loc[df['exclusion_flag'] == 'high-activity']

#checking output 
df_high.head()

df_high.shape

### 4.2 Excluding high-activity customers from the data and export the subset

#creating a new dataframe for just low-activity customers
df_low = df.loc[df['exclusion_flag'] == 'low-activity']

# checking output
df_low.head()

df_low.shape

# 5. Exporting newly created data sets

#exporting high activity data set
df_high.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'high_activity_customers.pkl'))

#exporting low activity data set
df_low.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'low_activity_customers.pkl'))
